*Update rAthena b4bdaebeeb

+++ src/map/elemental.c
@@ -569,7 +569,7 @@ int elemental_unlocktarget(struct elemental_data *ed) {
 }
 
 int elemental_skillnotok(uint16 skill_id, struct elemental_data *ed) {
-	int idx = skill_get_index(skill_id);
+	int16 idx = skill_get_index(skill_id);
 	nullpo_retr(1,ed);
 
 	if( idx == 0 )
@@ -580,7 +580,7 @@ int elemental_skillnotok(uint16 skill_id, struct elemental_data *ed) {
 
 struct skill_condition elemental_skill_get_requirements(uint16 skill_id, uint16 skill_lv){
 	struct skill_condition req;
-	int idx = skill_get_index(skill_id);
+	int16 idx = skill_get_index(skill_id);
 
 	memset(&req,0,sizeof(req));
